requirements.txt
discord.py
python-dotenv
requests
pillow
hashlib
shutil
cryptography